Fixes import loading error by renaming plugins.lua to load-plugins.lua
